requirements.txt
einops
FrEIA
getdist
hydra-core
matplotlib
numpy
pytorch
tensorboard
torchdiffeq